Move the SurveyLayout component to the PassSurvey feature
