+++ src/features/PassSurvey/ui/index.ts
@@ -1,2 +1,4 @@
 export * from './StartSurveyButton';
 export * from './SurveyManageButtons';
+export * from './ItemTimerBar';
+export { default as SurveyLayout } from './SurveyLayout';
